+++ android/Reguerta/app/src/test/java/com/reguerta/user/presentation/root/SessionCommunityActionsFailureTest.kt
@@ -2348,7 +2348,6 @@ private fun newsArticle(
     publishedBy = "Publisher",
     publishedAtMillis = 123L,
     urlImage = null,
-    publishedByUserId = "publisher-1",
 )
 
 private fun notificationEvent(