+++ ios/Reguerta/Reguerta/Data/News/InMemoryNewsRepository.swift
@@ -41,7 +41,6 @@ actor InMemoryNewsRepository: NewsRepository {
             body: article.body,
             active: article.active,
             publishedBy: article.publishedBy,
-            publishedByUserId: article.publishedByUserId,
             publishedAtMillis: article.publishedAtMillis,
             urlImage: article.urlImage
         )
@@ -105,7 +104,6 @@ private func localizedSeedNews(_ article: NewsArticle) async -> NewsArticle {
         body: body,
         active: article.active,
         publishedBy: article.publishedBy,
-        publishedByUserId: article.publishedByUserId,
         publishedAtMillis: article.publishedAtMillis,
         urlImage: article.urlImage
     )